+++ components/SearchField.tsx
@@ -3,12 +3,14 @@ import { Search } from "@equinor/eds-core-react";
 import { debounce } from "../utils/debounce";
 import React from "react";
 import styles from "./SearchField.module.scss";
+import removeEmpty from "utils/removeEmpty";
 
 export function SearchField(): JSX.Element {
   const router = useRouter();
   const handleSearch = (searchQuery: string) => {
-    router.query.searchQuery = searchQuery;
-    router.replace(router);
+    router.replace({
+      query: removeEmpty({ ...router.query, q: searchQuery }),
+    });
   };
 
   return (
@@ -17,7 +19,7 @@ export function SearchField(): JSX.Element {
       id="searchProjects"
       placeholder="Search by username or title"
       className={styles.searchField}
-      defaultValue={router.query.searchQuery}
+      defaultValue={router.query.q}
       onChange={(e) => {
         debounce(() => handleSearch(`${e.target.value}`), 500, "projectSearch");
       }}