+++ components/SideNavBarElement.tsx
@@ -10,7 +10,6 @@ export default function SideNavBarElement(props: {
 }): JSX.Element {
   const router = useRouter();
   const pathname = router.pathname;
-  const { searchQuery, orderBy } = router.query;
   const isOnCurrentPage = pathname == props.pathname;
 
   return (
@@ -21,7 +20,7 @@ export default function SideNavBarElement(props: {
         onClick={() =>
           router.push({
             pathname: props.pathname,
-            query: { orderBy, searchQuery },
+            query: router.query,
           })
         }
       >