+++ layouts/canvas.tsx
@@ -1,4 +1,3 @@
-import Head from "next/head";
 import {
   Button,
   Icon,
@@ -8,31 +7,39 @@ import {
   TopBar,
   Typography,
 } from "@equinor/eds-core-react";
-import { chevron_down, close, delete_forever, share } from "@equinor/eds-icons";
-import styles from "./default.layout.module.scss";
-import { useAccount, useIsAuthenticated, useMsal } from "@azure/msal-react";
 import React, { useEffect, useState } from "react";
-import UserMenu from "../components/AppHeader/UserMenu";
-import { useStoreDispatch, useStoreState } from "../hooks/storeHooks";
-import { useRouter } from "next/router";
-import BaseAPIServices from "../services/BaseAPIServices";
-import { HomeButton } from "./homeButton";
-import { RightTopBarSection } from "../components/RightTopBarSection";
-import { disableMouseWheelZoom } from "../utils/disableMouseWheelZoom";
-import { disableKeyboardZoomShortcuts } from "../utils/disableKeyboardZoomShortcuts";
-import { MySnackBar } from "../components/MySnackBar";
-import { AccessBox } from "../components/AccessBox";
-import { getMyAccess } from "../utils/getMyAccess";
+import { chevron_down, close, share } from "@equinor/eds-icons";
+import {
+  faveProject,
+  getProject,
+  unfaveProject,
+  updateProject,
+} from "../services/projectApi";
+import { useAccount, useIsAuthenticated, useMsal } from "@azure/msal-react";
 import { useMutation, useQuery, useQueryClient } from "react-query";
-import { getProject, updateProject } from "../services/projectApi";
-import { debounce } from "../utils/debounce";
-import { unknownErrorToString } from "../utils/isError";
-import packageJson from "../package.json";
-import { notifyOthers } from "../services/notifyOthers";
+import { useStoreDispatch, useStoreState } from "../hooks/storeHooks";
+
+import { AccessBox } from "../components/AccessBox";
+import BaseAPIServices from "../services/BaseAPIServices";
+import Head from "next/head";
+import Heart from "components/Heart";
+import { HomeButton } from "./homeButton";
+import { MySnackBar } from "../components/MySnackBar";
+import { RightTopBarSection } from "../components/RightTopBarSection";
 import { TooltipImproved } from "../components/TooltipImproved";
+import UserMenu from "../components/AppHeader/UserMenu";
+import { debounce } from "../utils/debounce";
+import { disableKeyboardZoomShortcuts } from "../utils/disableKeyboardZoomShortcuts";
+import { disableMouseWheelZoom } from "../utils/disableMouseWheelZoom";
+import { getMyAccess } from "../utils/getMyAccess";
+import { notifyOthers } from "../services/notifyOthers";
+import packageJson from "../package.json";
+import styles from "./default.layout.module.scss";
+import { unknownErrorToString } from "../utils/isError";
+import { useRouter } from "next/router";
 import { vsmProject } from "interfaces/VsmProject";
 
-const CanvasLayout = ({ children }) => {
+const CanvasLayout = ({ children }): JSX.Element => {
   const isAuthenticated = useIsAuthenticated();
 
   const router = useRouter();
@@ -51,6 +58,31 @@ const CanvasLayout = ({ children }) => {
         return queryClient.invalidateQueries();
       },
       onError: (e) => dispatch.setSnackMessage(unknownErrorToString(e)),
+    }
+  );
+
+  const [isMutatingFavourite, setIsMutatingFavourite] = useState(false);
+  const handleSettled = () => {
+    queryClient.invalidateQueries().then(() => setIsMutatingFavourite(false));
+  };
+
+  const faveMutation = useMutation(
+    () => {
+      setIsMutatingFavourite(true);
+      return faveProject(project?.vsmProjectID);
+    },
+    {
+      onSettled: handleSettled,
+    }
+  );
+
+  const unfaveMutation = useMutation(
+    () => {
+      setIsMutatingFavourite(true);
+      return unfaveProject(project?.vsmProjectID);
+    },
+    {
+      onSettled: handleSettled,
     }
   );
 
@@ -185,6 +217,14 @@ const CanvasLayout = ({ children }) => {
     );
   }
 
+  function handleDuplicate() {
+    if (project?.currentProcessId) {
+      router.push(`/process/${project.currentProcessId}/duplicate`);
+    } else {
+      router.push(`/process/${project.vsmProjectID}/duplicate`);
+    }
+  }
+
   return (
     <div style={{ overflow: "hidden" /* Hide scrollbars */ }}>
       <Head>
@@ -215,6 +255,12 @@ const CanvasLayout = ({ children }) => {
               >
                 <Icon data={chevron_down} title="chevron-down" size={16} />
               </Button>
+              <Heart
+                isFavourite={project?.isFavorite}
+                fave={() => faveMutation.mutate()}
+                unfave={() => unfaveMutation.mutate()}
+                isLoading={isMutatingFavourite}
+              />
             </div>
             <Menu
               id="menu-on-button"
@@ -240,11 +286,7 @@ const CanvasLayout = ({ children }) => {
                   Rename
                 </Typography>
               </Menu.Item>
-              <Menu.Item
-                onClick={() =>
-                  router.push(`/process/${project.vsmProjectID}/duplicate`)
-                }
-              >
+              <Menu.Item onClick={handleDuplicate}>
                 <Typography group="navigation" variant="menu_title" as="span">
                   Duplicate
                 </Typography>
@@ -350,13 +392,6 @@ const CanvasLayout = ({ children }) => {
               <>
                 <div className={styles.scrimHeaderWrapper}>
                   <div className={styles.scrimTitle}>Delete process</div>
-                  <Button
-                    autoFocus
-                    variant={"ghost_icon"}
-                    onClick={(e) => handleCloseDeleteScrim(e, false)}
-                  >
-                    <Icon data={close} title="Close" />
-                  </Button>
                 </div>
                 <div className={styles.scrimContent}>
                   {deleteError && (
@@ -364,22 +399,32 @@ const CanvasLayout = ({ children }) => {
                       {`${deleteError}`}
                     </Typography>
                   )}
-                  <Typography variant={"h4"}>
-                    Are you sure you want to delete the entire process?
-                  </Typography>
+                  <p>
+                    Are you sure you want to delete this process? By doing so
+                    you will delete all versions of Current and To-be processes,
+                    neither of which will be recoverable.
+                  </p>
                 </div>
-                <div className={styles.deleteButton}>
+                <div
+                  style={{
+                    display: "flex",
+                    justifyContent: "flex-end",
+                    gap: 12,
+                  }}
+                >
+                  <Button
+                    autoFocus
+                    variant={"outlined"}
+                    onClick={(e) => handleCloseDeleteScrim(e, false)}
+                  >
+                    Cancel
+                  </Button>
                   <Button
                     variant={"contained"}
                     color={"danger"}
                     onClick={() => deleteVSM()}
                   >
-                    <Icon
-                      data={delete_forever}
-                      title="Delete process"
-                      size={16}
-                    />
-                    Delete process
+                    Delete
                   </Button>
                 </div>
               </>